1. 修改Department model的id为Int 2. 学生和教师导入模板中院系信息为学院名称，而不是ID

+++ cloudapi-web/src/main/kotlin/cn/edu/buaa/scs/route/Course.kt
@@ -148,7 +148,7 @@ internal fun ApplicationCall.convertCourseResponse(course: Course, hasCount: Boo
         createTime = course.createTime,
         departmentId = course.departmentId,
         studentCnt = if (hasCount) this.course.studentCnt(course.id) else null,
-        departmentName = Department.id(course.departmentId).name
+        departmentName = Department.id(course.departmentId.toInt()).name
     )
 }
